+++ nautilus_trader/test_kit/rust/instruments.py
@@ -61,18 +61,24 @@ class TestInstrumentProviderPyo3:
         )
 
     @staticmethod
-    def btcusdt_future_binance(expiry: pd.Timestamp | None = None) -> CryptoFuture:
-        if expiry is None:
-            expiry = pd.Timestamp(datetime(2022, 3, 25), tz=pytz.UTC)
-            nanos_expiry = int(expiry.timestamp() * 1e9)
-        instrument_id_str = f"BTCUSDT_{expiry.strftime('%y%m%d')}.BINANCE"
+    def btcusdt_future_binance(
+        activation: pd.Timestamp | None = None,
+        expiration: pd.Timestamp | None = None,
+    ) -> CryptoFuture:
+        if activation is None:
+            activation = pd.Timestamp(2021, 12, 25, tz=pytz.utc)
+        if expiration is None:
+            expiration = pd.Timestamp(2022, 3, 25, tz=pytz.utc)
+
+        instrument_id_str = f"BTCUSDT_{expiration.strftime('%y%m%d')}.BINANCE"
         return CryptoFuture(  # type: ignore
             InstrumentId.from_str(instrument_id_str),
             Symbol("BTCUSDT"),
             TestTypesProviderPyo3.currency_btc(),
             TestTypesProviderPyo3.currency_usdt(),
             TestTypesProviderPyo3.currency_usdt(),
-            nanos_expiry,
+            activation.value,
+            expiration.value,
             2,
             6,
             Price.from_str("0.01"),
@@ -113,17 +119,22 @@ class TestInstrumentProviderPyo3:
         )
 
     @staticmethod
-    def appl_option(expiry: pd.Timestamp | None = None) -> OptionsContract:
-        if expiry is None:
-            expiry = pd.Timestamp(datetime(2021, 12, 17), tz=pytz.UTC)
-            nanos_expiry = int(expiry.timestamp() * 1e9)
+    def appl_option(
+        activation: pd.Timestamp | None = None,
+        expiration: pd.Timestamp | None = None,
+    ) -> OptionsContract:
+        if activation is None:
+            activation = pd.Timestamp(datetime(2021, 9, 17), tz=pytz.UTC)
+        if expiration is None:
+            expiration = pd.Timestamp(datetime(2021, 12, 17), tz=pytz.UTC)
         return OptionsContract(  # type: ignore
             InstrumentId.from_str("AAPL211217C00150000.OPRA"),
             Symbol("AAPL211217C00150000"),
             AssetClass.EQUITY,
             "AAPL",
             OptionKind.CALL,
-            nanos_expiry,
+            activation.value,
+            expiration.value,
             Price.from_str("149.0"),
             TestTypesProviderPyo3.currency_usdt(),
             2,
@@ -157,16 +168,21 @@ class TestInstrumentProviderPyo3:
         )
 
     @staticmethod
-    def futures_contract_es(expiry: pd.Timestamp | None = None) -> FuturesContract:
-        if expiry is None:
-            expiry = pd.Timestamp(datetime(2021, 12, 17), tz=pytz.UTC)
-            nanos_expiry = int(expiry.timestamp() * 1e9)
+    def futures_contract_es(
+        activation: pd.Timestamp | None = None,
+        expiration: pd.Timestamp | None = None,
+    ) -> FuturesContract:
+        if activation is None:
+            activation = pd.Timestamp(2021, 9, 17, tz=pytz.utc)
+        if expiration is None:
+            expiration = pd.Timestamp(2021, 12, 17, tz=pytz.utc)
         return FuturesContract(  # type: ignore
             InstrumentId.from_str("ESZ21.CME"),
             Symbol("ESZ21"),
             AssetClass.INDEX,
             "ES",
-            nanos_expiry,
+            activation.value,
+            expiration.value,
             TestTypesProviderPyo3.currency_usd(),
             2,
             Price.from_str("0.01"),